requirements.txt
pymupdf>=1.24.1
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
huggingface-hub>=0.23.0
streamlit>=1.35.0
numpy>=1.26.4
python-dotenv>=1.0.1
